+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # TelegramBotAPI changelog
 
+## 0.33.4
+
+* `Common`:
+    * `Version`:
+        * `uuid`: `0.2.3` -> `0.2.4`
+        * `MicroUtils`: `0.4.33` -> `0.4.35`
+* `Core`:
+    * All `TextSource` implementators have become `Serializable`
+        * New serializer `TextSourceSerializer`
+    * Interface`FromUserMessage` now extends `Message`
+    * New interface `FromUser`
+        * Interface `FromUserMessage` now extends `FromUser`
+* `Extensions Utils`
+    * Fixes in `parseCommandsWithParams`
+
 ## 0.33.3
 
 * `Common`: